chore(aient): Update submodule to latest commit
Updates the 'aient' submodule to point to a new commit. This synchronizes the project with the latest changes from the submodule repository.

The submodule has been advanced from commit ff23700 to 9d3fd47. Note that the new commit is marked as 'dirty', indicating there were uncommitted changes in the submodule's working directory.